Move existing types
